# update to new SFML version # various fixes # simplified build

+++ src/constants.hpp
@@ -16,7 +16,7 @@ namespace Constants{
 	const float c_playerBaseFadeFactor = 0.968f; //0.985f
 	const float c_projectileBaseSpeed = 10.0f;
 	const float c_projectileRadius = 5.f;
-	const float c_fastReloadCount = 7.f; // amount of ammo reloaded before it diminishes; only int values make sence
+	const int c_fastReloadCount = 7; // amount of ammo reloaded before it diminishes
 	const int	c_baseReloadSpeed = 40; // time for one fast reload projectile in frames
 	const float c_deathTreshhold = 0.05f;
